Refactor repo layout and dev dependencies
- de-nest SeqViewer and its container into one folder
- rename PartExplorer to SeqViewerContainer. Parts were specific to Loom's data model
- install packages so ESLint works again (was totally broken...)
- fix eslint and prettier dependencies
