+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,6 +1,6 @@
 ---
 name: Bug report
-about: Create a report to help us improve
+about: Create a report to help us improve SeqViz
 title: ""
 labels: ""
 assignees: ""
@@ -10,22 +10,14 @@ assignees: ""
 A description of what the bug is.
 
 **To Reproduce**
-Steps to reproduce the behavior:
+Steps to reproduce the behavior. Example:
 
-Example:
-
-```txt
 1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
-```
+2. Scroll down to '....'
+3. See error
 
 **Expected behavior**
 A description of what you expected to happen.
 
-**Screenshots**
-If applicable, add screenshots to help explain your problem.
-
 **Additional context**
-Add any other context about the problem here.
+Screenshots or any other context about the problem that helps describe it.
