+++ package.json
@@ -51,10 +51,18 @@
     "babel-plugin-transform-imports": "^2.0.0",
     "babel-polyfill": "^6.26.0",
     "babel-preset-react": "^6.24.1",
+    "eslint": "^6.7.2",
+    "eslint-config-prettier": "^6.7.0",
+    "eslint-plugin-import": "^2.19.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-react-hooks": "^1.7.0",
     "fs": "^0.0.2",
     "image-webpack-loader": "^6.0.0",
     "lodash-webpack-plugin": "^0.11.5",
     "node-sass": "^4.12.0",
+    "prettier": "^1.19.1",
     "react-scripts": "^3.0.1",
     "typescript": "^3.7.3",
     "uglifyjs-webpack-plugin": "^2.1.2",
