+++ src/SeqViewer/Circular/Circular.scss
