+++ src/SeqViewer/Circular/Labels/Labels.scss
