+++ src/SeqViewer/Linear/Linear.scss
