+++ src/SeqViewer/Linear/SeqBlock/Primers/Primers.jsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import shortid from "shortid";
-import { reverse } from "../../../../../utils/sequence";
+import { reverse } from "../../../../utils/sequence";
 
 /**
  * a single row of primers. Multiple of these may be in one seqBlock
