+++ src/SeqViewer/Linear/SeqBlock/SeqBlock.jsx
