+++ src/SeqViewer/SeqViewer.scss
