+++ src/SeqViewer/SeqViewerContainer.scss
