+++ src/SeqViewer/handlers/eventRouter.jsx
