+++ src/SeqViewer/handlers/index.js
