+++ src/SeqViewer/handlers/selectionHandler.jsx
@@ -1,7 +1,7 @@
 import { isEqual } from "lodash";
 import * as React from "react";
 
-import { calcGC, calcTm } from "../../../utils/sequence";
+import { calcGC, calcTm } from "../../utils/sequence";
 
 /**
  * an HOC dedicated to handling range selection for the LinearSeq viewer
